+++ Experiences/Experiences/View Controllers/CameraViewController.swift
@@ -109,35 +109,3 @@ class CameraViewController: UIViewController {
     }
     
 }
-
-extension CameraViewController: AVCaptureFileOutputRecordingDelegate {
-    
-    func fileOutput(_ output: AVCaptureFileOutput, didStartRecordingTo fileURL: URL, from connections: [AVCaptureConnection]) {
-        DispatchQueue.main.async {
-            self.updateButton()
-        }
-    }
-    
-    func fileOutput(_ output: AVCaptureFileOutput, didFinishRecordingTo outputFileURL: URL, from connections: [AVCaptureConnection], error: Error?) {
-        DispatchQueue.main.async {
-            defer { self.updateButton() }
-            
-            PHPhotoLibrary.requestAuthorization({ (status) in
-                if status != .authorized {
-                    NSLog("Please give VideoFilters access to your Photo Library in Settings.")
-                    return
-                }
-                
-                PHPhotoLibrary.shared().performChanges({
-                    PHAssetChangeRequest.creationRequestForAssetFromVideo(atFileURL: outputFileURL)
-                }, completionHandler: { (success, error) in
-                    if let error = error {
-                        NSLog("Error saving video to Photo Library: \(error.localizedDescription)")
-                        return
-                    }
-                })
-            })
-            
-        }
-    }
-}
